% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DC_crop}
\alias{DC_crop}
\title{Distance to coast}
\format{A RasterLayer with 50 m resolution. Distances are in meters.}
\usage{
DC_crop
}
\description{
Least cost distances from the coastline for the area of a dolphin survey.
}
\keyword{datasets}
