% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depth_crop}
\alias{depth_crop}
\title{Depth}
\format{A RasterLayer with 50 m resolution. Depth values are negative and in meters.}
\usage{
depth_crop
}
\description{
Depth for the area of a dolphin survey.
}
\keyword{datasets}
