% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DR_crop}
\alias{DR_crop}
\title{Distance to river}
\format{A RasterLayer with 50 m resolution. Distances are in meters.}
\usage{
DR_crop
}
\description{
Least cost distances from main rivers for the area of a dolphin survey.
}
\keyword{datasets}
