% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshm_boot.R
\name{dshm_boot}
\alias{dshm_boot}
\title{Non-parametric bootstrap for Hurdle model uncertainty}
\usage{
dshm_boot(det.fn.par, effects.pa = NULL, effects.ab = NULL, distdata,
  obsdata, segdata, model_fit, grid, group = FALSE, nsim, parallel = FALSE,
  ncores = NULL, mute = TRUE, stratification = "none")
}
\arguments{
\item{det.fn.par}{List of detection function parameters. For strucuture see the documentation for the 'ds' package.}

\item{effects.pa}{List of characters defining the binomial gam models to be fitted. For model structure see \code{\link[mgcv]{gam}}.}

\item{effects.ab}{List of characters defining the zero-truncated Poisson gam models to be fitted. For model structure see \code{\link[mgcv]{gam}}.}

\item{distdata}{Dataframe for distance sampling observations. For strucuture see the documentation for the 'ds' package.}

\item{obsdata}{Dataframe object containing 4 columns: (1) 'Sample.Label' (i.e. label for segments), (2) 'size' for cluster size, (3) 'distance' (in km) for perpendicular distance of sighting from the transect line, and (4) 'Effort' for segment length (in km).}

\item{segdata}{Dataframe object with at least 3 columns: (1) 'Transect.Label' (i.e. label for transects), (2) 'Sample.Label' (i.e. label for segments), and (3) 'Effort' for segment length (in km). It may also contain additional columns with relevant habitat covariates specific to each segment that will be fed into the spatial model.}

\item{model_fit}{Model fitted with the function 'dshm_fit'.}

\item{grid}{Grid used for model prediction. Column names for habitat covriates should correspond to those in 'segdata'.}

\item{group}{If TRUE group abundance is estimated (i.e. 'size' = 1).}

\item{nsim}{Number of simulations,}

\item{parallel}{If TRUE the simulations are performed on multiple cpus. Default is FALSE.}

\item{ncores}{Number of cpus for parallel execution.}

\item{mute}{If TRUE all unrelevant messages are suppressed. Default is TRUE.}

\item{stratification}{'transect' or 'stratum'}
}
\value{
A list of two arrays: (1) 'sim_grid' containing all simulated grids and (2) 'obs_fit' containing observed and fitted values for each simulation.
}
\description{
Non-parametric bootstrap for Hurdle model uncertainty
}
\author{
Filippo Franchini \email{filippo.franchini@outlook.com}
}
