% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshm_correct_segments.R
\name{dshm_correct_segments}
\alias{dshm_correct_segments}
\title{Corrects SpatialPolygons for overlapping features}
\usage{
dshm_correct_segments(data, intersections)
}
\arguments{
\item{data}{SpatialPolygonsDataFrame.}

\item{intersections}{List for all segment intersections.}
}
\value{
Polygons corrected for overlapping features.
}
\description{
Corrects SpatialPolygons for overlapping features
}
\author{
Filippo Franchini \email{filippo.franchini@outlook.com}
}
