% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshm_make_grid.R
\name{dshm_make_grid}
\alias{dshm_make_grid}
\title{Creates a grid}
\usage{
dshm_make_grid(extent, cell.size, projection)
}
\arguments{
\item{extent}{Grid extent as specified in raster.}

\item{cell.size}{Grid cell size in meters.}

\item{projection}{Grid CRS as specified in raster.}
}
\description{
Creates a grid
}
