% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshm_plot.R
\name{dshm_plot}
\alias{dshm_plot}
\title{Plotting prediction grid}
\usage{
dshm_plot(prediction, grid, probability = FALSE, sightings = NULL,
  plot_title = NULL, scale_col = c("blue", "yellow", "red"),
  scale_lim = NULL, scale_val = NULL, cex = 1, saveRaster = FALSE,
  raster_name = NULL, resolution)
}
\arguments{
\item{prediction}{Predictions for each grid cell in the prediction grid.}

\item{grid}{Grid as SpatialPolygonsDataFrame.}

\item{probability}{If \code{TRUE} colour label is Pr(Presence), if \code{FALSE} is Abundance. Default is \code{FALSE}.}

\item{sightings}{sighting dataframe with at least three columns:
\itemize{
  \item x: sighting x coordinate.
  \item y: sighting y coordinate.
  \item size: sighting size.
}}

\item{plot_title}{Title of the plot.}

\item{scale_col}{Vector with colour scale values. Length should correspond to the length of 'scale_val'.}

\item{scale_lim}{Vector for lower and upper limits of the colour scale.}

\item{scale_val}{Vector for the scale values corresponding to colour transition as defined in 'scale_col'.}

\item{cex}{Size of the pixels in the map (visualization purpose only, for true scaled spatial map save as raster).}

\item{saveRaster}{If \code{TRUE} the map is saved as a raster file. Default is \code{TRUE}.}

\item{raster_name}{Name of the saved raster.}

\item{resolution}{Raster resolution.}
}
\value{
Prediction map and raster image (if \code{saveRaster = TRUE}).
}
\description{
\code{dshm_plot} plots prediction grids, makes easier changing scale colours and intervals, provides the possibility to save the grid as a .shp file.
}
\details{
For more information about fitting Hurdle models, plotting model predictions and much more you can download the \href{http://github.com/FilippoFranchini/dshm/blob/master/vignettes}{fitting_Hurldle.pdf} tutorial.
}
\author{
Filippo Franchini \email{filippo.franchini@outlook.com}
}
