% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land_crop}
\alias{land_crop}
\title{Land}
\format{A SpatialPolygonsDataFrame with 30 features (i.e. islands). Associated data are not important.}
\usage{
land_crop
}
\description{
Land for the area of the dolphin survey.
}
\keyword{datasets}
