% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obsdata}
\alias{obsdata}
\title{Observation data}
\format{A data frame with 16 rows and 10 variables:
\describe{
  \item{Region.Label}{ID for stratum}
  \item{Area}{Stratum area in km2}
  \item{Transect.Label}{ID for transect}
  \item{Sample.Label}{ID for segment}
  \item{Latitude}{Segment centroid latitude}
  \item{Longitude}{Segment centroid longitude}
  \item{distance}{Sighting perpendicular distance from the transect line}
  \item{size}{Number of dolphin}
  \item{object}{Sighting ID}
  \item{detected}{Required for \code{\link[Distance]{ds}} function}
}}
\usage{
obsdata
}
\description{
A dataset containing relevant information for dolphin sightings.
}
\keyword{datasets}
