% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strata}
\alias{strata}
\title{Strata}
\format{A SpatialPolygonsDataFrame with 9 features (i.e. strata). Except for ID, associated data are not important.
#' \describe{
  \item{ID}{ID for stratum}
}}
\usage{
strata
}
\description{
Strata covered by a boat during a dolphin survey.
}
\keyword{datasets}
