% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transects}
\alias{transects}
\title{Transects}
\format{A SpatialLinesDataFrame with 15 features (i.e. lines). Associated data are not important.}
\usage{
transects
}
\description{
Transect lines covered by a boat during a dolphin survey.
}
\keyword{datasets}
