% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshm_diagnostics.R
\name{dshm_diagnostics}
\alias{dshm_diagnostics}
\title{Computes CDF and plots CDF vs. EDF relationship for dshm objects}
\usage{
dshm_diagnostics(fit.pa, fit.ab, obs, mute = FALSE, plot = TRUE,
  plot.n = 1)
}
\arguments{
\item{fit.pa}{Fitted values from binomial submodel}

\item{fit.ab}{Fitted values from zero-truncated Poisson submodel}

\item{obs}{Observation (same length as fit.pa and fit.ab)}

\item{mute}{If TRUE returns p-value and test statistics for Kolmogorov-Smirnov test. Default is FALSE.}

\item{plot}{If TRUE plots CDF vs EDF and fitted vs. observed values plots. Default is TRUE.}

\item{plot.n}{1 for CDF vs EDF plot and 2 for both CDF vs EDF and fitted vs. observed values plots.}
}
\description{
Computes CDF and plots CDF vs. EDF relationship for dshm objects
}
