% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshm_fill_grid.R
\name{dshm_fill_grid}
\alias{dshm_fill_grid}
\title{Corrects grid shapefile for land and calculates covariate statistics within each grid cell}
\usage{
dshm_fill_grid(empty.grid, land.data = NULL, cov, fun, ncores = 2)
}
\arguments{
\item{empty.grid}{Empty grid shapefile.}

\item{land.data}{Coastline shapefile.}

\item{cov}{List of covriate rasters.}

\item{fun}{Function for covariate statistcs within each cell (i.e. mean or median).}

\item{ncores}{Number of cores (this function is run in parallel only).}
}
\description{
Corrects grid shapefile for land and calculates covariate statistics within each grid cell
}
