% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshm_finalize_segments.R
\name{dshm_finalize_segments}
\alias{dshm_finalize_segments}
\title{Corrects segment shapefile for land and calculates covariate statistics for each segement.}
\usage{
dshm_finalize_segments(segment.data, land.data = NULL, covariates, fun,
  parallel = FALSE, ncores = NULL)
}
\arguments{
\item{segment.data}{Segment shapefile.}

\item{land.data}{Coastline shapefile.}

\item{covariates}{List of covariate raters.}

\item{fun}{Function to calculate the covariate statistics within each segment.}

\item{parallel}{If TRUE the function is run on multiple cores. Default is FALSE.}

\item{ncores}{Number of cores if parallel is TRUE.}
}
\description{
Corrects segment shapefile for land and calculates covariate statistics for each segement.
}
