% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshm_fit.R
\name{dshm_fit}
\alias{dshm_fit}
\title{Fits and averages spatial Hurdle models}
\usage{
dshm_fit(det.fn, effects.pa = NULL, effects.ab = NULL, knots.pa = NULL,
  knots.ab = NULL, method = "GCV.Cp", lim = 0.1, obsdata, segdata, grid,
  SelectionTable = TRUE, showSelectedModels = FALSE, group = FALSE,
  strip.width = NULL)
}
\arguments{
\item{det.fn}{Detection function fitted by \code{\link[Distance]{ds}}.}

\item{effects.pa}{List of characters defining the binomial gam models to be fitted. For model structure see \code{\link[mgcv]{gam}}.}

\item{effects.ab}{List of characters defining the zero-truncated Poisson gam models to be fitted. For model structure see \code{\link[mgcv]{gam}}.}

\item{knots.pa}{List of knot gam knot positions for each smooth term of the fitted binomial models.}

\item{knots.ab}{List of knot gam knot positions for each smooth term of the fitted zero-truncated Poisson models.}

\item{method}{Character. GAM fitting method. Note that 'REML' is not available since it is not campatible with \code{\link[countreg]{ztpoisson}}. Default is 'GCV.Cp'.}

\item{lim}{AIC weight (AICw) threshold for model averaging. Models with AICw < lim are not averaged. Default is 0.1.}

\item{obsdata}{Dataframe object containing 4 columns: (1) 'Sample.Label' (i.e. label for segments), (2) 'size' for cluster size, (3) 'distance' (in km) for perpendicular distance of sighting from the transect line, and (4) 'Effort' for segment length (in km).}

\item{segdata}{Dataframe object with at least 3 columns: (1) 'Transect.Label' (i.e. label for transects), (2) 'Sample.Label' (i.e. label for segments), and (3) 'Effort' for segment length (in km). It may also contain additional columns with relevant habitat covariates specific to each segment that will be fed into the spatial model.}

\item{grid}{Grid used for model prediction. Column names for habitat covriates should correspond to those in 'segdata'.}

\item{SelectionTable}{If TRUE model selection table is reported for each submodel. Default is TRUE.}

\item{showSelectedModels}{If TRUE best fitted submodel variants are reported. Default is FALSE.}

\item{group}{If TRUE group abundance is estimated (i.e. 'size' = 1).}

\item{strip.width}{Strip width to calculate segment area if there is no "area" column.}
}
\value{
A list of 3 objects: (1) 'models' containing a list for all presence-absence submodel variants ('pa') and all submodel variants for abundance conditional on presence ('ab'); (2) 'info' containing a list of variant weights and best variant ids for both submodels ('weights.pa' and 'weights.ab', and 'ID.pa' and 'ID.ab' respectively); and (3) 'grid_data' containing a list of prediction grids for presence-absence submodel ('pa'), submodel for abundance conditional on presence ('ab'), and Hurdle model ('H').
}
\description{
Fits and averages spatial Hurdle models
}
\author{
Filippo Franchini \email{filippo.franchini@outlook.com}
}
