% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshm_plot.R
\name{dshm_plot}
\alias{dshm_plot}
\title{Plots prediction grid and saves it as a raster image}
\usage{
dshm_plot(prediction, species = NULL, grid, map_type = "Hurdle",
  sightings = FALSE, sightingData, scale_col = c("blue", "yellow", "red"),
  scale_lim = NULL, scale_val = NULL, cex = 1, saveRaster = FALSE,
  grid.shp)
}
\arguments{
\item{prediction}{Prediction grid returned by dshm_fit$grid_data.}

\item{species}{Character. Species name (used for plotting and saving the raster).}

\item{grid}{Grid used for model prediction. Column names for habitat covriates should correspond to those in 'segdata'.}

\item{map_type}{Character. Type of map to plot: 'presence_absence', 'abundance', 'Hurdle', or 'SD'.}

\item{sightings}{If TRUE, sightings are plotted in the map. Default is FALSE.}

\item{sightingData}{Dataframe with at least three columns: 'x' and 'y' coordinates and 'size' for loaction and size of each sighitng, respectively.}

\item{scale_col}{Vector with colour scale values. Length should correspond to the length of 'scale_val'.}

\item{scale_lim}{Vector for lower and upper limits of the colour scale.}

\item{scale_val}{Vector for the scale values corresponding to colour transition as defined in 'scale_col'.}

\item{cex}{Size of the pixels in the map (visualization purpose only, for true scaled spatial map save as raster).}

\item{saveRaster}{If TRUE the map is saved as a raster file. A grid shapefile with projection is needed. Default is FALSE.}

\item{grid.shp}{Grid shapefile (SpatialPolygonsDataFrame object) with specific projection.}
}
\value{
Map and raster. Raster name contains information in 'map_type' and 'species'.
}
\description{
Plots prediction grid and saves it as a raster image
}
\author{
Filippo Franchini \email{filippo.franchini@outlook.com}
}
