% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshm_spatial_correlation.R
\name{dshm_spatial_correlation}
\alias{dshm_spatial_correlation}
\title{Calculates spatial correlation for dshm}
\usage{
dshm_spatial_correlation(coord, z, xlab, ylab, lims)
}
\arguments{
\item{coord}{Centroids of each segment.}

\item{z}{Fitted values of model residuals.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{lims}{A vector of minimum and maximum correlation values.}
}
\description{
Calculates spatial correlation for dshm
}
