% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshm_split_transects.R
\name{dshm_split_transects}
\alias{dshm_split_transects}
\title{Splits transect lines into segments and adds buffer as well as metadata}
\usage{
dshm_split_transects(transect.data, inter.dist, lwr, search.time, w,
  parallel = FALSE, ncores = NULL, cap)
}
\arguments{
\item{transect.data}{Transect lines as SpatialLinesDataFrame.}

\item{inter.dist}{Distance between each segment in meters.}

\item{lwr}{Lower limit in meters for segments.}

\item{search.time}{Time the routine has to search for each set of cutting points.}

\item{w}{Buffer in meters.}

\item{parallel}{If TRUE the rountine is carried out by multiple cores. Default is FALSE.}

\item{ncores}{Number of cores.}

\item{cap}{If TRUE a cap is added to the ends of each splitted transect.}
}
\value{
Splitted transect lines with buffer and information about: (1) 'Transect.Label' the original transect id, (2) 'Sample.Label' the segment id, (3) 'lenght' the length of the segment, and (4) 'area' the area of the segment.
}
\description{
Splits transect lines into segments and adds buffer as well as metadata
}
\author{
Filippo Franchini \email{filippo.franchini@outlook.com}
}
